Revert "Merge branch 'libroDiario' into develop"
This reverts commit bedb6ab619702ddc543a12f7b6178beee9043250, reversing
changes made to e6dd2a5fd08938d8ed14eea83335ccc0f9ded91d.

+++ Models/Cuentas.java
@@ -12,7 +12,7 @@ public class Cuentas {
     private String codigo;
     private String nombre;
     private int id_tipo_saldo;
-    private float saldo;
+    private long saldo;
     
     public Cuentas(){
        
@@ -49,7 +49,7 @@ public class Cuentas {
         this.id_tipo_saldo = id_tipo_saldo;
     }
 
-    public float getSaldo() {
+    public long getSaldo() {
         return saldo;
     }
